refactor(tokens): removing tokens/upload
