+++ packages/components/upload/index.ts
@@ -8,3 +8,4 @@ export * from './src/upload'
 export * from './src/upload-content'
 export * from './src/upload-list'
 export * from './src/upload-dragger'
+export * from './src/constants'
